split rules based on os
