+++ client/boxoptions/rules_other.go
@@ -1,0 +1,9 @@
+//go:build !linux
+
+package boxoptions
+
+import "github.com/sagernet/sing-box/option"
+
+func Options() option.Options {
+	return boxOptions
+}
